Calculate hash now uses picosha
Be aware that this may or may not be valid code, as picosha may not be able to hash the transactions. 
In this case, we will have to accumulate the id of each transaction, and then hash those together. 
OR we can resort to the merkle tree, which may be easier at this point.